Remove future dependency

+++ test/integration/clitest/test_metadata.py
@@ -20,7 +20,6 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 import pytest
-from past.builtins import long
 
 import omero
 import omero.gateway
@@ -43,7 +42,7 @@ class MetadataTestBase(CLITest):
 
         conn = BlitzGateway(client_obj=self.client)
         self.imageid = unwrap(self.image.getId())
-        assert isinstance(self.imageid, long)
+        assert isinstance(self.imageid, int)
         wrapper = conn.getObject("Image", self.imageid)
         self.md = Metadata(wrapper)
